반차 근무 로직 개선 (오전/오후 반차별 출근 가능 시간 및 지각 기준 분리)
- MORNING_HALF: 13시 이후 출근 허용, 13시 기준 지각 판단
- AFTERNOON_HALF: 오전 근무만 허용, 9시 기준 지각 판단
- AttendanceServiceImpl clockIn() 반차 상태 반영

+++ src/main/java/org/goodee/startup_BE/approval/dto/ApprovalDocRequestDTO.java
@@ -46,7 +46,7 @@ public class ApprovalDocRequestDTO {
 
     // 휴가(AT1)
     private Long vacationTypeCode;   // CommonCode ID
-    private Integer vacationDays;
+    private Double vacationDays;
     private String vacationReason;
 
     // 출장(AT2)
